+++ .gitignore
@@ -3,4 +3,4 @@
 *.import
 .import
 *.o
-android
+demo/android
