+++ README.md
@@ -11,20 +11,13 @@ How to use
 ----------
 - Configure, install  and enable the "Android Custom Template" for your project, just follow the [official documentation](https://docs.godotengine.org/en/latest/getting_started/workflow/export/android_custom_build.html);
 - go to the [release tab](https://github.com/Shin-NiL/Godot-Android-Admob-Plugin/releases), choose a version and download the respective zip package;
-- extract the ```admob-plugin``` directory (from the zip package) inside the ```res://android/``` directory on your Godot project.
+- extract the content of the ```admob-plugin``` directory (from the zip package) inside the ```res://android/plugins``` directory on your Godot project.
 - extract the ```admob-lib``` directory (from the zip package) inside the ```res://``` directory on your Godot project.
-- on the Project -> Export -> Android -> Options -> Permissions: check the permissions for _Access Network State_ and _Internet_
-- on the Project Settings -> Android -> Modules, add the string:
-
-```
-org/godotengine/godot/GodotAdMob
-```
-- Edit the file ```res//android/build/gradle.properties``` addind the following lines at the end:
-```
-android.useAndroidX=true
-android.enableJetifier=true
-```
-- Edit the file ```res//android/admob-plugin/AndroidManifest.conf``` changing the sample id (_ca-app-pub-3940256099942544~3347511713_) for your _APPLICATION ID_ on the line
+- on the Project -> Export... -> Android -> Options -> 
+    - Permissions: check the permissions for _Access Network State_ and _Internet_
+    - Custom Template: check the _Use Custom Build_
+    - Plugins: check the _Godot Ad Mob_ (this plugin)
+- Edit the file ```res//android/plugins/AndroidManifest.conf``` changing the sample id (_ca-app-pub-3940256099942544~3347511713_) for your _APPLICATION ID_ on the line
 ```
 android:value="ca-app-pub-3940256099942544~3347511713"/
 ```
@@ -194,9 +187,9 @@ Troubleshooting
 ```
 adb logcat -s godot
 ```
-* _Admob Java Singleton not found_: this module is Android only, so the Admob Java singleton will only exists on Android platform. In other words, you will be able to run it on an Android device (or emulator) only, it will not work on editor or on another platform.
+* _Admob Java Singleton not found_: this module is Android only, so the Admob Java singleton will only exists on the Android platform. In other words, you will be able to run it on an Android device (or emulator) only, it will not work on editor or on another platform.
 
-* Error code 3 (_ERROR_CODE_NO_FILL_) is a common issue with Admob, but out of the scope to this module. Here's the description on the API page: [ERROR_CODE_NO_FILL: The ad request was successful, but no ad was returned due to lack of ad inventory.](https://developers.google.com/android/reference/com/google/android/gms/ads/AdRequest.html#ERROR_CODE_NO_FILL)
+* Error code 3 (_ERROR_CODE_NO_FILL_) is a common issue with Admob, but out of the scope to this module. Here's the description on the API page: [ERROR_CODE_NO_FILL: The ad request was successful, but no ad was returned due to lack of ad inventory.](https://developers.google.com/android/reference/com/google/android/gms/ads/AdRequest.html#ERROR_CODE_NO_FILL) You can find more information about the error codes [here](https://support.google.com/admob/thread/3494603).
 
 References
 -------------
