--- admob-plugin/AndroidManifest.conf
@@ -1,5 +1,0 @@
-[application]
-	<!-- Sample AdMob App ID: ca-app-pub-3940256099942544~3347511713 -->
-	<meta-data
-		android:name="com.google.android.gms.ads.APPLICATION_ID"
-		android:value="ca-app-pub-3940256099942544~3347511713"/>
