+++ admob-plugin/godot-lib.release/.gitignore
@@ -1,0 +1,1 @@
+/build
