+++ admob-plugin/godot-lib.release/build.gradle
@@ -1,0 +1,2 @@
+configurations.maybeCreate("default")
+artifacts.add("default", file('godot-lib.release.aar'))
