+++ admob-plugin/godotadmob/src/androidTest/java/shinnil/godot/plugin/android/godotadmob/ExampleInstrumentedTest.java
@@ -1,0 +1,27 @@
+package shinnil.godot.plugin.android.godotadmob;
+
+import android.content.Context;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
+
+import static org.junit.Assert.assertEquals;
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+ */
+@RunWith(AndroidJUnit4.class)
+public class ExampleInstrumentedTest {
+    @Test
+    public void useAppContext() {
+        // Context of the app under test.
+        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
+
+        assertEquals("shinnil.godot.plugin.android.godotadmob.test", appContext.getPackageName());
+    }
+}
