+++ admob-plugin/godotadmob/src/main/java/shinnil/godot/plugin/android/godotadmob/Banner.java
@@ -1,41 +1,39 @@
-package org.godotengine.godot;
+package shinnil.godot.plugin.android.godotadmob;
 
 import android.app.Activity;
-import com.google.android.gms.ads.MobileAds;
-import com.google.android.gms.ads.AdRequest;
-import com.google.android.gms.ads.AdListener;
-import com.google.android.gms.ads.AdView;
-import com.google.android.gms.ads.AdSize;
-import android.widget.FrameLayout;
-import android.view.Gravity;
 import android.graphics.Color;
-import android.view.View;
-
 import android.util.Log;
-import org.godotengine.godot.*;
+import android.view.Gravity;
+import android.view.View;
+import android.widget.FrameLayout;
+
+import com.google.android.gms.ads.AdListener;
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.AdSize;
+import com.google.android.gms.ads.AdView;
+
+import org.godotengine.godot.GodotLib;
 
 public class Banner {
     private AdView adView = null; // Banner view
     private FrameLayout layout = null;
-    private FrameLayout.LayoutParams adParams = null; 
+    private FrameLayout.LayoutParams adParams = null;
     private AdRequest adRequest = null;
     private Activity activity = null;
-   
-	public Banner(final String id, final AdRequest adRequest, final Activity activity, final int instanceId, final boolean isOnTop, final FrameLayout layout)
-	{
+
+    public Banner(final String id, final AdRequest adRequest, final Activity activity, final int instanceId, final boolean isOnTop, final FrameLayout layout) {
         this.activity = activity;
         this.layout = layout;
         this.adRequest = adRequest;
 
         adParams = new FrameLayout.LayoutParams(
-            FrameLayout.LayoutParams.MATCH_PARENT,
-            FrameLayout.LayoutParams.WRAP_CONTENT
+                FrameLayout.LayoutParams.MATCH_PARENT,
+                FrameLayout.LayoutParams.WRAP_CONTENT
         );
-        if(isOnTop) adParams.gravity = Gravity.TOP;
+        if (isOnTop) adParams.gravity = Gravity.TOP;
         else adParams.gravity = Gravity.BOTTOM;
 
-        if (adView != null)
-        {
+        if (adView != null) {
             layout.removeView(adView); // Remove the old view
         }
 
@@ -45,19 +43,17 @@ public class Banner {
         adView.setBackgroundColor(Color.TRANSPARENT);
 
         adView.setAdSize(AdSize.SMART_BANNER);
-        adView.setAdListener(new AdListener()
-        {
+        adView.setAdListener(new AdListener() {
             @Override
             public void onAdLoaded() {
                 Log.w("godot", "AdMob: onAdLoaded");
-                GodotLib.calldeferred(instanceId, "_on_admob_ad_loaded", new Object[]{ });
+                GodotLib.calldeferred(instanceId, "_on_admob_ad_loaded", new Object[]{});
             }
 
             @Override
-            public void onAdFailedToLoad(int errorCode)
-            {
+            public void onAdFailedToLoad(int errorCode) {
                 Log.w("godot", "AdMob: onAdFailedToLoad. errorCode: " + errorCode);
-                GodotLib.calldeferred(instanceId, "_on_admob_banner_failed_to_load", new Object[] { errorCode });
+                GodotLib.calldeferred(instanceId, "_on_admob_banner_failed_to_load", new Object[]{errorCode});
             }
         });
         layout.addView(adView, adParams);
@@ -66,12 +62,11 @@ public class Banner {
         adView.loadAd(adRequest);
 
     }
-    
-	public void show()
-	{
+
+    public void show() {
         if (adView == null) {
             Log.w("w", "AdMob: showBanner - banner not loaded");
-            return;	
+            return;
         }
 
         if (adView.getVisibility() == View.VISIBLE) {
@@ -83,10 +78,8 @@ public class Banner {
         Log.d("godot", "AdMob: Show Banner");
     }
 
-	public void resize()
-	{
-        if (layout == null || adView == null || adParams == null)
-        {
+    public void resize() {
+        if (layout == null || adView == null || adParams == null) {
             return;
         }
 
@@ -96,8 +89,8 @@ public class Banner {
 
         int gravity = adParams.gravity;
         adParams = new FrameLayout.LayoutParams(
-            FrameLayout.LayoutParams.MATCH_PARENT,
-            FrameLayout.LayoutParams.WRAP_CONTENT
+                FrameLayout.LayoutParams.MATCH_PARENT,
+                FrameLayout.LayoutParams.WRAP_CONTENT
         );
         adParams.gravity = gravity;
         AdListener adListener = adView.getAdListener();
@@ -120,24 +113,20 @@ public class Banner {
     }
 
 
-	public void hide()
-	{
+    public void hide() {
         if (adView.getVisibility() == View.GONE) return;
         adView.setVisibility(View.GONE);
         adView.pause();
         Log.d("godot", "AdMob: Hide Banner");
     }
-    
-	public int getWidth()
-	{
+
+    public int getWidth() {
         return AdSize.SMART_BANNER.getWidthInPixels(activity);
     }
-    
-	public int getHeight()
-	{
+
+    public int getHeight() {
         return AdSize.SMART_BANNER.getHeightInPixels(activity);
-	}
+    }
 
 
-    
 }
