+++ admob-plugin/godotadmob/src/main/java/shinnil/godot/plugin/android/godotadmob/Interstitial.java
@@ -1,32 +1,31 @@
-package org.godotengine.godot;
+package shinnil.godot.plugin.android.godotadmob;
 
 import android.app.Activity;
-import com.google.android.gms.ads.MobileAds;
+import android.util.Log;
+
+import com.google.android.gms.ads.AdListener;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.InterstitialAd;
-import com.google.android.gms.ads.AdListener;
 
-import android.util.Log;
-import org.godotengine.godot.*;
+import org.godotengine.godot.GodotLib;
 
 public class Interstitial {
     private InterstitialAd interstitialAd = null; // Interstitial object
-    
+
     public Interstitial(final String id, final AdRequest adRequest, final Activity activity, final int instanceId) {
         interstitialAd = new InterstitialAd(activity);
         interstitialAd.setAdUnitId(id);
-        interstitialAd.setAdListener(new AdListener()
-        {
+        interstitialAd.setAdListener(new AdListener() {
             @Override
             public void onAdLoaded() {
                 Log.w("godot", "AdMob: onAdLoaded");
-                GodotLib.calldeferred(instanceId, "_on_interstitial_loaded", new Object[] { });
+                GodotLib.calldeferred(instanceId, "_on_interstitial_loaded", new Object[]{});
             }
 
             @Override
             public void onAdFailedToLoad(int errorCode) {
                 Log.w("godot", "AdMob: onAdFailedToLoad(int errorCode) - error code: " + Integer.toString(errorCode));
-                GodotLib.calldeferred(instanceId, "_on_interstitial_failed_to_load", new Object[] { errorCode });
+                GodotLib.calldeferred(instanceId, "_on_interstitial_failed_to_load", new Object[]{errorCode});
             }
 
             @Override
@@ -41,7 +40,7 @@ public class Interstitial {
 
             @Override
             public void onAdClosed() {
-                GodotLib.calldeferred(instanceId, "_on_interstitial_close", new Object[] { });
+                GodotLib.calldeferred(instanceId, "_on_interstitial_close", new Object[]{});
                 interstitialAd.loadAd(adRequest);
                 Log.w("godot", "AdMob: onAdClosed");
             }
