+++ admob-plugin/godotadmob/src/main/java/shinnil/godot/plugin/android/godotadmob/RewardedVideo.java
@@ -1,14 +1,15 @@
-package org.godotengine.godot;
+package shinnil.godot.plugin.android.godotadmob;
 
 import android.app.Activity;
+import android.util.Log;
+
+import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.MobileAds;
 import com.google.android.gms.ads.reward.RewardItem;
 import com.google.android.gms.ads.reward.RewardedVideoAd;
 import com.google.android.gms.ads.reward.RewardedVideoAdListener;
-import com.google.android.gms.ads.AdRequest;
 
-import android.util.Log;
-import org.godotengine.godot.*;
+import org.godotengine.godot.GodotLib;
 
 
 public class RewardedVideo {
@@ -21,31 +22,31 @@ public class RewardedVideo {
             @Override
             public void onRewardedVideoAdLeftApplication() {
                 Log.w("godot", "AdMob: onRewardedVideoAdLeftApplication");
-                GodotLib.calldeferred(instanceId, "_on_rewarded_video_ad_left_application", new Object[] {});
+                GodotLib.calldeferred(instanceId, "_on_rewarded_video_ad_left_application", new Object[]{});
             }
 
             @Override
             public void onRewardedVideoAdClosed() {
                 Log.w("godot", "AdMob: onRewardedVideoAdClosed");
-                GodotLib.calldeferred(instanceId, "_on_rewarded_video_ad_closed", new Object[] {});
+                GodotLib.calldeferred(instanceId, "_on_rewarded_video_ad_closed", new Object[]{});
             }
 
             @Override
             public void onRewardedVideoAdFailedToLoad(int errorCode) {
                 Log.w("godot", "AdMob: onRewardedVideoAdFailedToLoad. errorCode: " + errorCode);
-                GodotLib.calldeferred(instanceId, "_on_rewarded_video_ad_failed_to_load", new Object[] { errorCode });
+                GodotLib.calldeferred(instanceId, "_on_rewarded_video_ad_failed_to_load", new Object[]{errorCode});
             }
 
             @Override
             public void onRewardedVideoAdLoaded() {
                 Log.w("godot", "AdMob: onRewardedVideoAdLoaded");
-                GodotLib.calldeferred(instanceId, "_on_rewarded_video_ad_loaded", new Object[] {});
+                GodotLib.calldeferred(instanceId, "_on_rewarded_video_ad_loaded", new Object[]{});
             }
 
             @Override
             public void onRewardedVideoAdOpened() {
                 Log.w("godot", "AdMob: onRewardedVideoAdOpened");
-                GodotLib.calldeferred(instanceId, "_on_rewarded_video_ad_opened", new Object[] {});
+                GodotLib.calldeferred(instanceId, "_on_rewarded_video_ad_opened", new Object[]{});
             }
 
             @Override
@@ -53,19 +54,19 @@ public class RewardedVideo {
                 Log.w("godot", "AdMob: "
                         + String.format(" onRewarded! currency: %s amount: %d", reward.getType(), reward.getAmount()));
                 GodotLib.calldeferred(instanceId, "_on_rewarded",
-                        new Object[] { reward.getType(), reward.getAmount() });
+                        new Object[]{reward.getType(), reward.getAmount()});
             }
 
             @Override
             public void onRewardedVideoStarted() {
                 Log.w("godot", "AdMob: onRewardedVideoStarted");
-                GodotLib.calldeferred(instanceId, "_on_rewarded_video_started", new Object[] {});
+                GodotLib.calldeferred(instanceId, "_on_rewarded_video_started", new Object[]{});
             }
 
             @Override
             public void onRewardedVideoCompleted() {
                 Log.w("godot", "AdMob: onRewardedVideoCompleted");
-                GodotLib.calldeferred(instanceId, "_on_rewarded_video_completed", new Object[] {});
+                GodotLib.calldeferred(instanceId, "_on_rewarded_video_completed", new Object[]{});
             }
         });
     }
