--- admob-plugin/gradle.conf
@@ -1,2 +1,0 @@
-[dependencies]
-    api ('com.google.android.gms:play-services-ads:19.1.0')
