+++ admob-plugin/settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name='GodotAdmob'
+include ':app', ':godot-lib.release'
+include ':godotadmob'
