+++ demo/export_presets.cfg
@@ -22,6 +22,8 @@ one_click_deploy/clear_previous_install=true
 custom_template/debug=""
 custom_template/release=""
 custom_template/use_custom_build=true
+plugins/GodotPayment=false
+plugins/GodotAdMob=true
 command_line/extra_args=""
 version/code=1
 version/name="1.0"
@@ -48,7 +50,7 @@ apk_expansion/enable=false
 apk_expansion/SALT=""
 apk_expansion/public_key=""
 architectures/armeabi-v7a=true
-architectures/arm64-v8a=true
+architectures/arm64-v8a=false
 architectures/x86=false
 architectures/x86_64=false
 permissions/custom_permissions=PoolStringArray(  )
